services/feature-pipeline/requirements.txt
kafka-python==2.0.2
redis==5.0.1
numpy==1.26.3
